feat: Improve `useChainState` and standardize contract loading (#50)
### Description

It improves the chain state helper definition and loading and exposes a `useContracts` hook used throughout the app.
The problem with this PR right now is that the addresses in the SDK are not the ones that are indexed in `The Graph` that is a separate deployment. I will migrate the indexers to this deployment as well.

After I do that, however, we will have no test proposals because the ones that I created are on the other deployment.
We need a small script to seed the deployment with some test proposals that have votes. 

### Other changes

- None

### Tested

- Verified correct contract addresses were being used with logs on chain switch :)

### Related issues

-  Fixes #49